[Sooraj] [tweetable/#6] Moving method from helper to model
What changes you have made?
Moved query_passages_by_status method to model from controller

Why it is needed?
The query of a model should reside inside model and it is not helpers
job to query a model

How does it impacts?
The method won’t be available in helper anymore